Modified file location & formatting - Moved all deepseek related features/APIs into a folder named 'deepSeek'
