+++ src/backend/deepSeek/DeepSeekFeature3.js
@@ -1,5 +1,5 @@
-const { queryDeepSeekV3 } = require('./deepseek.js');
-const prompts = require('./prompts/deepseekPrompts.js');
+const { queryDeepSeekV3 } = require('./APIdeepseek.js');
+const prompts = require('../prompts/deepseekPrompts.js');
 
 
 async function getSummary(data){
